Move report related functions into a separate file
