+++ index.js
@@ -6,7 +6,6 @@ import {omit, isEqual} from 'lodash-es';
 import micromatch from 'micromatch';
 import arrify from 'arrify';
 import pMap from 'p-map';
-import defineLazyProperty from 'define-lazy-prop';
 import slash from 'slash';
 import {
 	normalizeOptions,
@@ -15,72 +14,7 @@ import {
 	buildConfig,
 	mergeOptions,
 } from './lib/options-manager.js';
-
-/** Merge multiple reports into a single report */
-const mergeReports = reports => {
-	const report = {
-		results: [],
-		errorCount: 0,
-		warningCount: 0,
-	};
-
-	for (const currentReport of reports) {
-		report.results.push(...currentReport.results);
-		report.errorCount += currentReport.errorCount;
-		report.warningCount += currentReport.warningCount;
-	}
-
-	return report;
-};
-
-const getReportStatistics = results => {
-	const statistics = {
-		errorCount: 0,
-		warningCount: 0,
-		fixableErrorCount: 0,
-		fixableWarningCount: 0,
-	};
-
-	for (const result of results) {
-		statistics.errorCount += result.errorCount;
-		statistics.warningCount += result.warningCount;
-		statistics.fixableErrorCount += result.fixableErrorCount;
-		statistics.fixableWarningCount += result.fixableWarningCount;
-	}
-
-	return statistics;
-};
-
-const processReport = (report, {isQuiet = false} = {}) => {
-	if (isQuiet) {
-		report = ESLint.getErrorResults(report);
-	}
-
-	const result = {
-		results: report,
-		...getReportStatistics(report),
-	};
-
-	defineLazyProperty(result, 'usedDeprecatedRules', () => {
-		const seenRules = new Set();
-		const rules = [];
-
-		for (const {usedDeprecatedRules} of report) {
-			for (const rule of usedDeprecatedRules) {
-				if (seenRules.has(rule.ruleId)) {
-					continue;
-				}
-
-				seenRules.add(rule.ruleId);
-				rules.push(rule);
-			}
-		}
-
-		return rules;
-	});
-
-	return result;
-};
+import {mergeReports, processReport} from './lib/report.js';
 
 const runEslint = async (filePath, options, processorOptions) => {
 	const engine = new ESLint(omit(options, ['filePath', 'warnIgnored']));
